+++ FreeRTOS/Source/portable/IAR/ARM_CM3/portmacro.h
@@ -145,7 +145,7 @@ extern void vPortYield( void );
 	/*-----------------------------------------------------------*/
 
 	#include <intrinsics.h>
-	#define portGET_HIGHEST_PRIORITY( uxTopPriority, uxReadyPriorities ) uxTopPriority = ( 31 - __CLZ( ( uxReadyPriorities ) ) )
+	#define portGET_HIGHEST_PRIORITY( uxTopPriority, uxReadyPriorities ) uxTopPriority = ( 31UL - ( ( uint32_t ) __CLZ( ( uxReadyPriorities ) ) ) )
 
 #endif /* configUSE_PORT_OPTIMISED_TASK_SELECTION */
 /*-----------------------------------------------------------*/
